# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dosa Ji Restaurant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  customer: 'customer',
  shopkeeper: 'shopkeeper',
  admin: 'admin',
} as const;

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  shopkeeper: 'shopkeeper',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  imageUrl: string;
  isVeg: boolean;
  isAvailable: boolean;
  isFeatured: boolean;
  createdAt: string;
}

export interface MenuItemInput {
  /** @minLength 2 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  category: string;
  imageUrl: string;
  isVeg: boolean;
  isAvailable?: boolean;
  isFeatured?: boolean;
}

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  imageUrl?: string;
  isVeg?: boolean;
  isAvailable?: boolean;
  isFeatured?: boolean;
}

export interface AvailabilityUpdate {
  isAvailable: boolean;
}

export interface OrderItem {
  menuItemId: number;
  name: string;
  price: number;
  quantity: number;
  imageUrl: string;
}

export interface DeliveryAddress {
  name: string;
  phone: string;
  addressLine1: string;
  /** @nullable */
  addressLine2?: string | null;
  city: string;
  pincode: string;
}

export interface OrderInput {
  items: OrderItem[];
  deliveryAddress: DeliveryAddress;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  razorpayOrderId?: string | null;
  /** @nullable */
  razorpayPaymentId?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  Placed: 'Placed',
  Preparing: 'Preparing',
  OutForDelivery: 'OutForDelivery',
  Delivered: 'Delivered',
  Cancelled: 'Cancelled',
} as const;

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  customerName?: string | null;
  items: OrderItem[];
  deliveryAddress: DeliveryAddress;
  subtotal: number;
  deliveryCharge: number;
  gst: number;
  couponDiscount: number;
  total: number;
  /** @nullable */
  razorpayOrderId?: string | null;
  /** @nullable */
  razorpayPaymentId?: string | null;
  status: OrderStatus;
  createdAt: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  Placed: 'Placed',
  Preparing: 'Preparing',
  OutForDelivery: 'OutForDelivery',
  Delivered: 'Delivered',
  Cancelled: 'Cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface PaymentOrderInput {
  amount: number;
}

export interface PaymentOrderResponse {
  id: string;
  amount: number;
  currency: string;
  /** @nullable */
  keyId?: string | null;
}

export interface PaymentVerifyInput {
  orderId: number;
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
}

export type CouponDiscountType = typeof CouponDiscountType[keyof typeof CouponDiscountType];


export const CouponDiscountType = {
  percent: 'percent',
  flat: 'flat',
} as const;

export interface Coupon {
  id: number;
  code: string;
  discountType: CouponDiscountType;
  value: number;
  minOrder: number;
  maxUses: number;
  usedCount: number;
  /** @nullable */
  expiresAt?: string | null;
  isActive: boolean;
  createdAt: string;
}

export type CouponInputDiscountType = typeof CouponInputDiscountType[keyof typeof CouponInputDiscountType];


export const CouponInputDiscountType = {
  percent: 'percent',
  flat: 'flat',
} as const;

export interface CouponInput {
  code: string;
  discountType: CouponInputDiscountType;
  value: number;
  minOrder: number;
  maxUses: number;
  /** @nullable */
  expiresAt?: string | null;
  isActive?: boolean;
}

export type CouponUpdateDiscountType = typeof CouponUpdateDiscountType[keyof typeof CouponUpdateDiscountType];


export const CouponUpdateDiscountType = {
  percent: 'percent',
  flat: 'flat',
} as const;

export interface CouponUpdate {
  code?: string;
  discountType?: CouponUpdateDiscountType;
  value?: number;
  minOrder?: number;
  maxUses?: number;
  /** @nullable */
  expiresAt?: string | null;
  isActive?: boolean;
}

export interface CouponApplyInput {
  code: string;
  orderTotal: number;
}

export interface CouponApplyResult {
  discount: number;
  finalTotal: number;
  message: string;
}

export type UserRoleUpdateRole = typeof UserRoleUpdateRole[keyof typeof UserRoleUpdateRole];


export const UserRoleUpdateRole = {
  customer: 'customer',
  shopkeeper: 'shopkeeper',
  admin: 'admin',
} as const;

export interface UserRoleUpdate {
  role: UserRoleUpdateRole;
}

export interface UserStatusUpdate {
  isActive: boolean;
}

export interface DashboardStats {
  totalMenuItems: number;
  todayOrders: number;
  todayRevenue: number;
  pendingOrders: number;
  totalOrders: number;
  totalUsers: number;
  totalRevenue: number;
}

export interface RevenueDataPoint {
  date: string;
  revenue: number;
}

export interface OrderDayPoint {
  day: string;
  count: number;
}

export type ListMenuItemsParams = {
category?: string;
isVeg?: boolean;
featured?: boolean;
};

export type ListAllOrdersParams = {
status?: string;
limit?: number;
};