# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dosa Ji Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── MENU ────────────────────────────────────────────────────────────────────
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: isVeg
          in: query
          schema:
            type: boolean
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item (shopkeeper/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item (shopkeeper/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /menu/{id}/availability:
    patch:
      operationId: toggleMenuItemAvailability
      tags: [menu]
      summary: Toggle availability (shopkeeper/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"

  # ── ORDERS ───────────────────────────────────────────────────────────────────
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List coupons (admin)
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"

    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create a coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/apply:
    post:
      operationId: applyCoupon
      tags: [coupons]
      summary: Validate and apply a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponApplyInput"
      responses:
        "200":
          description: Coupon applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponApplyResult"
        "400":
          description: Invalid coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update a coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"

    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete a coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── USERS ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── AUTH ──────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [customer, shopkeeper, admin]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, shopkeeper, admin]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string

    # ── MENU ──────────────────────────────────────────────────────────────────
    MenuItem:
      type: object
      required: [id, name, description, price, category, imageUrl, isVeg, isAvailable, isFeatured, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        isVeg:
          type: boolean
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          type: string
    MenuItemInput:
      type: object
      required: [name, description, price, category, imageUrl, isVeg]
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        price:
          type: number
          minimum: 0
        category:
          type: string
        imageUrl:
          type: string
        isVeg:
          type: boolean
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
    MenuItemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        isVeg:
          type: boolean
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
    AvailabilityUpdate:
      type: object
      required: [isAvailable]
      properties:
        isAvailable:
          type: boolean

    # ── ORDERS ────────────────────────────────────────────────────────────────
...
    Coupon:
      type: object
      required: [id, code, discountType, value, minOrder, maxUses, usedCount, isActive, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
          enum: [percent, flat]
        value:
          type: number
        minOrder:
          type: number
        maxUses:
          type: integer
        usedCount:
          type: integer
        expiresAt:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    CouponInput:
      type: object
      required: [code, discountType, value, minOrder, maxUses]
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percent, flat]
        value:
          type: number
        minOrder:
          type: number
        maxUses:
          type: integer
        expiresAt:
          type: ["string", "null"]
        isActive:
          type: boolean
    CouponUpdate:
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percent, flat]
        value:
          type: number
        minOrder:
          type: number
        maxUses:
          type: integer
        expiresAt:
          type: ["string", "null"]
        isActive:
          type: boolean
    CouponApplyInput:
      type: object
      required: [code, orderTotal]
      properties:
        code:
          type: string
        orderTotal:
          type: number
    CouponApplyResult:
      type: object
      required: [discount, finalTotal, message]
      properties:
        discount:
          type: number
        finalTotal:
          type: number
        message:
          type: string

    # ── USERS ─────────────────────────────────────────────────────────────────
...