# lib/api-zod/src/generated/types/coupon.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dosa Ji Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { CouponDiscountType } from './couponDiscountType';

export interface Coupon {
  id: number;
  code: string;
  discountType: CouponDiscountType;
  value: number;
  minOrder: number;
  maxUses: number;
  usedCount: number;
  /** @nullable */
  expiresAt?: string | null;
  isActive: boolean;
  createdAt: string;
}