# lib/api-zod/src/generated/types/couponApplyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dosa Ji Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface CouponApplyInput {
  code: string;
  orderTotal: number;
}